tweak record adding to DB
