+++ WolvenKit.RED4.TweakDB/TweakDB.cs
@@ -1,6 +1,9 @@
+using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Runtime.InteropServices;
 using WolvenKit.Core.Extensions;
+using WolvenKit.RED4.TweakDB.Serialization;
 using WolvenKit.RED4.TweakDB.Types;
 
 namespace WolvenKit.RED4.TweakDB
@@ -26,11 +29,12 @@ namespace WolvenKit.RED4.TweakDB
     /// </remarks>
     public class TweakDB
     {
-        private static readonly uint s_magic = 0xBB1DB47;
-        private static readonly uint s_blobVersion = 5;
-        private static readonly uint s_parserVersion = 4;
+        private const uint s_magic = 0xBB1DB47;
+        private const uint s_blobVersion = 5;
+        private const uint s_parserVersion = 4;
 
         private readonly FlatsPool _flats = new();
+        private readonly Dictionary<string, string> _records = new();
 
         /// <summary>
         /// Add a new flat value to the pool.
@@ -38,6 +42,24 @@ namespace WolvenKit.RED4.TweakDB
         /// <param name="name">The flat's name.</param>
         /// <param name="value">The value.</param>
         public void Add(string name, IType value) => _flats.Add(name, value);
+
+        /// <summary>
+        /// Add a new record to the pool.
+        /// </summary>
+        /// <param name="name">The flat's name.</param>
+        /// <param name="record">The value.</param>
+        public void Add(string name, TweakRecord record)
+        {
+            foreach (var (key, flat) in record.Members)
+            {
+                _flats.Add(key, flat);
+            }
+
+            var recordParent = record.Inherits; //unused
+            var recordType = record.Type;
+
+            _records.Add(name, recordType);
+        }
 
         /// <summary>
         /// Save the database to a file.
