--- WolvenKit.RED4.TweakDB/TweakFlatDto.cs
@@ -1,6 +1,0 @@
-using WolvenKit.RED4.TweakDB.Types;
-
-namespace WolvenKit.RED4.TweakDB
-{
-    
-}
